backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
django-filter==25.2
requests==2.31.0
djangorestframework-simplejwt==5.5.0
